refactor: extract PDF region block list assignment
